move simple-transformers to transformers
