+++ src/churn-pipe/churn-pipe.ts
@@ -3,7 +3,7 @@
 /// <reference path='../../../third_party/typings/es6-promise/es6-promise.d.ts' />
 
 import aqm = require('../aqm/aqm');
-import caesar = require('../simple-transformers/caesar');
+import caesar = require('../transformers/caesar');
 import churn_types = require('../churn/churn.types');
 import decompression = require('../fancy-transformers/decompressionShaper');
 import encryption = require('../fancy-transformers/encryptionShaper');
@@ -11,7 +11,7 @@ import fragmentation = require('../fancy-transformers/fragmentationShaper');
 import ipaddr = require('ipaddr.js');
 import logging = require('../logging/logging');
 import net = require('../net/net.types');
-import PassThrough = require('../simple-transformers/passthrough');
+import PassThrough = require('../transformers/passthrough');
 import protean = require('../fancy-transformers/protean');
 import sequence = require('../fancy-transformers/byteSequenceShaper');
